Makefile and Workflow Updates for Dev1
Why these changes are being introduced:
Lambda functions that we manage are integrated across across a couple of
infrastructure repositories. The mitlib-tf-workloads-ecr repository
builds the ECR infrastructure for the container and generates the
necessary GitHub Actions workflow files and the Makefile commands for
building and publishing the container image to AWS.

How this addresses that need:
* Create the dev_build.yml file for the automated build/publish of the
container to Dev1
* Update the Makefile with the developer build/publish commands for
manually testing changes in Dev1

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/ENSY-177