gets export file using job ID
Why these changes are being introduced:
We need the exact name of the export file in order to retrieve it from from s3.

The job end webhook data does not include all of the information necessary to derive the
filename.

The filename takes the following format [alma bursar transfer job name]-[job id]-[timestamp].xml. The job end
webhook includes only the job name, and job id, but not the timestamp.

 How this addresses that need:
* Uses the path-like SOURCE_PREFIX (from environment variable)
job name, and job id (both passed in from step function) as a prefix
to filter a list of objects in the source bucket. Job Ids are unique,
so there should be only one object that matches.

 Side effects of this change:
* lambda should be invoked with a dict like
{"job_name": "job name", "job_id" : "job id"}