Create Stage and Prod Automation Workflows
Why these changes are being introduced:
With the changes from mitlib-tf-workloads-ecr pushed to Dev1,
Stage-Workloads, and Prod-Workloads, we can finish the GitHub Actions
workflows for this repository.

How this addresses that need:
* Create the stage_build workflow for automated build/push of the
container to ECR (triggered by commits to the `main` branch)
* Create the prod_promote workflow for automated download from stage
and upload to prod ECR (triggered by tagged release on the `main`
branch)

Side effects of this change:
None. Other than the fact that every commit to the `main` branch will
now automatically update the container image in the Stage-Workloads
ECR repository

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/ENSY-177