return total charges and record count
When uploading the .csv to CHASS, the operator is required to fill in form fields with the
total of all the charges being submitted, and the number of records in the input file.

* calculates total charges and record count
* returns these values in the dict which the step function can access
so they can be included in the success email.

Step function email needs to be updated to take advantage of
these new values

* https://mitlibraries.atlassian.net/browse/ENSY-187